refactor(api-core): move filteredOrganizations logic inside organizations resource
